🐛  Type conversion
Type(s): FIX

+++ app/Http/Controllers/Partner/OperationController.php
@@ -254,7 +254,7 @@ class OperationController extends Controller
             'category_id' => $secondary_category->id
         ];
 
-        if ($partner->package_id === config('sheba.partner_lite_packages_id')) $current_category_partner['is_partner_premise_applied'] = true;
+        if ($partner->package_id === (int)config('sheba.partner_lite_packages_id')) $current_category_partner['is_partner_premise_applied'] = true;
 
         $secondary_category->load(['services' => function ($q) use ($location) {
             $q->whereExists(function ($query) use ($location) {
